framework/media: sync from ecode Media
sync Media framework till below commit of Media:
https://github.ecodesamsung.com/TizenRT/Media/commit/7a79583e10e24e3f1715329582016da79f193279

+++ framework/src/media/stream_info.c
@@ -30,7 +30,7 @@ int stream_info_create(stream_policy_t stream_policy, stream_info_t **stream_inf
 	if (*stream_info == NULL) {
 		return -ENOMEM;
 	}
-	(*stream_info)->id = (uint64_t)(*stream_info);
+	(*stream_info)->id = (size_t)(*stream_info);
 	(*stream_info)->policy = stream_policy;
 
 	return OK;
